+++ lib/laflem/collections/python/virtualenv/__init__.py
@@ -3,3 +3,5 @@ Import all modules in this package.
 """
 from .cleanup import VirtualenvCleanupModule
 from .create import VirtualenvCreateModule
+from .list import VirtualenvListModule
+from .delete import VirtualenvDeleteModule